Feat: Calculator 클래스 내에서 실제 사칙연산 로직 구현

+++ src/Calculator/domain/Operation.java
@@ -1,33 +1,31 @@
 package Calculator.domain;
-import java.util.Scanner;
-import java.util.function.BiFunction;
 
 //실제 사칙연산 수행
 public enum Operation {
 
     PLUS("+") {
         @Override
-        public double calculate(int a, int b) {
-            return (double) a + b;
+        public double calculate(double a, double b) {
+            return  a + b;
         }
     },
     MINUS("-") {
         @Override
-        public double calculate(int a, int b) {
-            return (double) a - b;
+        public double calculate(double a, double b) {
+            return  a - b;
         }
     },
     MULTIPLY("*") {
         @Override
-        public double calculate(int a, int b) {
-            return (double) a * b;
+        public double calculate(double a, double b) {
+            return a * b;
         }
     },
     DIVIDE("/") {
         @Override
-        public double calculate(int a, int b) {
+        public double calculate(double a, double b) {
             // 0으로 나누는 예외 처리는 apply()에서 처리합니다.
-            return (double) a / b;
+            return a / b;
         }
     };
 
@@ -39,9 +37,9 @@ public enum Operation {
 
     //추상 메서드 정의 (abstract)
     //추상메서드가 없으면 PLUS,MINUS 같은 Operation Enum 상수는 각각의 고유한 계산 로직을 오버라이딩(구현)할수가 없음
-    public abstract double calculate(int a, int b);
+    public abstract double calculate(double a, double b);
 
-    public double apply(int a, int b){
+    public double apply(double a, double b){
         return calculate(a,b);
     }
